requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Caching
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
#redis==5.0.1

# Background Tasks
celery[redis]==5.3.4

# Web3 & Blockchain
web3==6.12.0
eth-account==0.9.0
eth-utils==2.3.1

# HTTP Clients & APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Twitter Integration
tweepy==4.14.0

# Data Processing
pandas==2.1.4
numpy==1.26.0

# Validation & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

pydantic-settings
PyJWT
flower
psutil
celery[redis]
